+++ models/yolo.py
@@ -135,8 +135,6 @@ class IDetect(nn.Module):
                 y[..., 0:2] = (y[..., 0:2] * 2. - 0.5 + self.grid[i]) * self.stride[i]  # xy
                 y[..., 2:4] = (y[..., 2:4] * 2) ** 2 * self.anchor_grid[i]  # wh
                 z.append(y.view(bs, -1, self.no))
-        # return x
-        # return z
         return x if self.training else (torch.cat(z, 1), x)
     
     def fuseforward(self, x):
