+++ test.py
@@ -57,88 +57,7 @@ def test(data,
         (save_dir / 'labels' if save_txt else save_dir).mkdir(parents=True, exist_ok=True)  # make dir
 
         # Load model
-        model = attempt_load(weights,load_mode='train',fuse=False, map_location=device)  # load FP32 model
-        import torch.nn as nn
-        class detectConfig(nn.Module):
-            def __init__(self,na,nc,nl,anchors,stride,anchor_grid, grid, no):
-                super(detectConfig, self).__init__()
-                self.na = na
-                self.nc = nc
-                self.nl = nl
-                self.anchors = anchors
-                self.stride = stride
-                self.anchor_grid = anchor_grid
-                self.grid = grid
-                self.no = no
-            @staticmethod
-            def _make_grid(nx=20, ny=20):
-                yv, xv = torch.meshgrid([torch.arange(ny), torch.arange(nx)])
-                return torch.stack((xv, yv), 2).view((1, 1, ny, nx, 2)).float()
-            
-            def forward(self, x):
-                z = []
-                for i in range(self.nl):
-                    bs, _, ny, nx, _  = x[i].shape
-                    # print(x[i].shape)
-                    if self.grid[i].shape[2:4] != x[i].shape[2:4]:
-                        self.grid[i] = self._make_grid(nx, ny).to(x[i].device)
-                    y = x[i].sigmoid()
-                    y[..., 0:2] = (y[..., 0:2] * 2. - 0.5 + self.grid[i]) * self.stride[i]  # xy
-                    y[..., 2:4] = (y[..., 2:4] * 2) ** 2 * self.anchor_grid[i]  # wh
-
-                    z.append(y.view(bs, -1, self.no))      
-                
-                return (torch.cat(z,1), x)
-        # save configurations
-        stride = model.stride
-        names = model.names
-        idetect = model.model[-1]
-        dconfig = detectConfig(na=idetect.na, nc=idetect.nc, nl=idetect.nl, anchors=idetect.anchors, stride=idetect.stride,
-                                    anchor_grid=idetect.anchor_grid, grid=idetect.grid, no=idetect.no)
-        if opt.graphmodule:
-            import torch.nn as nn
-            class detectConfig(nn.Module):
-                def __init__(self,na,nc,nl,anchors,stride,anchor_grid, grid, no):
-                    super(detectConfig, self).__init__()
-                    self.na = na
-                    self.nc = nc
-                    self.nl = nl
-                    self.anchors = anchors
-                    self.stride = stride
-                    self.anchor_grid = anchor_grid
-                    self.grid = grid
-                    self.no = no
-                @staticmethod
-                def _make_grid(nx=20, ny=20):
-                    yv, xv = torch.meshgrid([torch.arange(ny), torch.arange(nx)])
-                    return torch.stack((xv, yv), 2).view((1, 1, ny, nx, 2)).float()
-                
-                def forward(self, x):
-                    z = []
-                    for i in range(self.nl):
-                        bs, _, ny, nx, _  = x[i].shape
-                        # print(x[i].shape)
-                        if self.grid[i].shape[2:4] != x[i].shape[2:4]:
-                            self.grid[i] = self._make_grid(nx, ny).to(x[i].device)
-                        y = x[i].sigmoid()
-                        y[..., 0:2] = (y[..., 0:2] * 2. - 0.5 + self.grid[i]) * self.stride[i]  # xy
-                        y[..., 2:4] = (y[..., 2:4] * 2) ** 2 * self.anchor_grid[i]  # wh
-
-                        z.append(y.view(bs, -1, self.no))      
-                    
-                    return (torch.cat(z,1), x)
-            # save configurations
-            stride = model.stride
-            names = model.names
-            idetect = model.model[-1]
-            dconfig = detectConfig(na=idetect.na, nc=idetect.nc, nl=idetect.nl, anchors=idetect.anchors, stride=idetect.stride,
-                                        anchor_grid=idetect.anchor_grid, grid=idetect.grid, no=idetect.no)
-            # load graphmodule model
-            model = torch.load(opt.graphmodule, map_location=device)
-            # attach configurations
-            model.stride = stride
-            model.names = names
-            model.model = nn.Sequential(dconfig)
+        model = attempt_load(weights,load_mode='train',graphmodule=opt.graphmodule, fuse=False, map_location=device)  # load FP32 model
             
         gs = max(int(model.stride.max()), 32)  # grid size (max stride)
         imgsz = check_img_size(imgsz, s=gs)  # check img_size
@@ -152,7 +71,7 @@ def test(data,
         model.half()
 
     # Configure
-    # model.eval()
+    model.eval()
     if isinstance(data, str):
         is_coco = data.endswith('coco.yaml')
         with open(data) as f:
@@ -196,9 +115,9 @@ def test(data,
             # Run model
             t = time_synchronized()
             if opt.graphmodule:
-                out, train_out = dconfig(model(img))  # inference and training outputs
+                out, train_out = model.model[-1](model(img))  # inference and training outputs
             else:
-                out, train_out = dconfig(model(img, augment=augment))  # inference and training outputs
+                out, train_out = model(img, augment=augment)  # inference and training outputs
             t0 += time_synchronized() - t
 
             # Compute loss
